Computer vision seems to work
Within camera frame, we use four ArUco markers as the base (coordinates known), and we put the fifth marker on the moving speaker. With homography in openCV, the position in camera were mathched with real coordinates setting. The 'real' coordinate of fifth moving maker is then calculated and out put in excel form.